third_party/WebKit/LayoutTests/fast/css/invalidation/indeterminate-pseudo.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
:indeterminate { background-color: green }
input + div { color: pink }
</style>
<input id="checkbox" type="checkbox"></input>
<div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</div>
<script>
description("Use descendant invalidation set for :indeterminate pseudo class.")

var transparent = "rgba(0, 0, 0, 0)";
var green = "rgb(0, 128, 0)";

shouldBe("getComputedStyle(checkbox, '').backgroundColor", "transparent");

checkbox.offsetTop; // Force recalc.

checkbox.indeterminate = true;

if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");

shouldBe("getComputedStyle(checkbox, '').backgroundColor", "green");
</script>
